backend/requirements.txt
# Core Framework
Django==6.0
djangorestframework==3.15.2

# Database
psycopg2-binary==2.9.10

# Background Tasks (Celery)
celery==5.4.0
redis==5.0.1

# HTTP Requests
requests==2.31.0

# Optional: Monitoring
# flower==2.0.1  # Web UI for monitoring Celery workers
# sentry-sdk==1.40.0  # Error tracking
